+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Allow safe swapping of USB stick, including creating API endpoint for getting and setting shots folder path
 - Show storage status under disk usage graph, which checks for write access, and which displays subdirectories of `/media` otherwise
 - Add dropdown button to download app and motion log files to settings page
+- Introduce setting `trigger light`
 
 ### Changed