+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Show storage status under disk usage graph, which checks for write access, and which displays subdirectories of `/media` otherwise
 - Add dropdown button to download app and motion log files to settings page
 - Introduce setting `trigger light`
+- Introduce working hours settings
 
 ### Changed